SCRUM-34 add unit test for card
