+++ src/__tests__/card.test.tsx
@@ -1,4 +1,8 @@
-import { fetch_figure_cards } from "@/lib/card";
+import { Card } from "@/components/cards";
+import { fetch_figure_cards, GetMoveCard } from "@/lib/card";
+import { render, screen } from "@testing-library/react";
+import "@testing-library/jest-dom";
+
 import fetchMock from "jest-fetch-mock";
 
 fetchMock.enableMocks();
@@ -46,4 +50,49 @@ describe("fetch_figure_cards", () => {
 
     expect(result).toEqual(mockResponse);
   });
+});
+
+describe("GetMoveCard", () => {
+  beforeEach(() => {
+    fetchMock.resetMocks();
+  });
+  test('url should be correct', async () => {
+    const result = await GetMoveCard({ player_id: 1 });
+    console.log(result);
+    expect(fetch).toHaveBeenCalledWith('http://localhost:8000/movement-cards', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ player_id: 1 }),
+    });
+  });
+
+  test('should return success when the fetch request is successful', async () => {
+    const mockResponse = { status: 'OK', message: 'Movement cards fetched successfully' };
+    fetchMock.mockResponseOnce(JSON.stringify(mockResponse), { status: 200 });
+    const result = await GetMoveCard({ player_id: 1 });
+    expect(result).toEqual(mockResponse);
+  });
+
+  test('should return error when the fetch request fails', async () => {
+    const mockResponse = { status: 'ERROR', message: 'An error occurred while getting the movement cards' };
+    fetchMock.mockRejectOnce(new Error('Network error'));
+    const result = await GetMoveCard({ player_id: 1 });
+    expect(result).toEqual(mockResponse);
+  });
+});
+
+
+describe('Card Component', () => {
+  test('renders correctly', () => {
+    render(<Card link="c0" />);
+    const cardElement = screen.getByRole('img', { name: /c0/i });
+    expect(cardElement).toBeInTheDocument();
+  });
+
+  test('has the correct alt text', () => {
+    render(<Card link="c0" />);
+    const cardElement = screen.getByRole('img', { name: /c0/i });
+    expect(cardElement).toHaveAttribute('alt', 'c0');
+  });
+
 });
